refactor: Сервисы работы юзера перенесены в другой пакет
